requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
arrow==1.1.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==1.1.1
multidict==5.1.0
nb-cli==0.4.1
nonebot-adapter-cqhttp==2.0.0a13
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a13.post1
poyo==0.5.0
prompt-toolkit==1.0.14
pydantic==1.8.1
pyfiglet==0.8.post1
Pygments==2.8.1
pygtrie==2.4.2
PyInquirer==1.0.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==5.0.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.13.6
text-unidecode==1.3
tomlkit==0.7.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
win32-setctime==1.0.3
yarl==1.6.3
